requirements.txt
python-telegram-bot==20.5
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
